adjust skill modifier structure
